[struct pack] Detect hash conflict in debug mode
