+++ src/struct_pack/tests/CMakeLists.txt
@@ -1,6 +1,8 @@
 set(CMAKE_RUNTIME_OUTPUT_DIRECTORY ${CMAKE_BINARY_DIR}/tests)
 add_executable(test_serialize
         test_serialize.cpp
+        test_compile_time_calculate.cpp
+        test_data_struct.cpp
         test_tuplet.cpp
         test_alignas.cpp
         test_pragma_pack.cpp
@@ -8,5 +10,5 @@ add_executable(test_serialize
         main.cpp
         )
 add_test(NAME test_serialize COMMAND test_serialize)
-target_compile_definitions(test_serialize PRIVATE STRUCT_PACK_USE_INT32_SIZE STRUCT_PACK_ENABLE_UNPORTABLE_TYPE)
+target_compile_definitions(test_serialize PRIVATE STRUCT_PACK_ENABLE_UNPORTABLE_TYPE)
 target_link_libraries(test_serialize PRIVATE libstruct_pack doctest)
